Finished v1.0-rc1 of LawlJSON
Improved the interface for LJValue to be safer and cleaner to use.
Made it so that the Serializer for LJStrings properly converts values into their escaped equivalents.
Added a parsing system that accepts a JSON string and returns parsed data.

The library still has not been tested much so be careful if you choose to use this release.